refactor: rename package from com.mtlog.goland to com.mtlog.analyzer
  - Fix plugin ID validation error by removing "goland" from ID
  - Update all package declarations and imports throughout codebase
  - Move all source files to new package directory structure
  - Update plugin.xml and test descriptors with new class names
  - Remove continue-on-error from plugin verifier in CI
