+++ goland-plugin/src/main/resources/META-INF/plugin.xml
@@ -1,5 +1,5 @@
 <idea-plugin>
-    <id>com.mtlog.goland</id>
+    <id>com.mtlog.analyzer</id>
     <name>mtlog-analyzer</name>
     <vendor email="support@mtlog.com" url="https://github.com/willibrandon/mtlog">mtlog</vendor>
     
@@ -13,11 +13,11 @@
     <extensions defaultExtensionNs="com.intellij">
         <externalAnnotator 
             language="go"
-            implementationClass="com.mtlog.goland.annotator.MtlogExternalAnnotator"/>
+            implementationClass="com.mtlog.analyzer.annotator.MtlogExternalAnnotator"/>
             
         <localInspection 
             language="go"
-            implementationClass="com.mtlog.goland.inspection.MtlogBatchInspection"
+            implementationClass="com.mtlog.analyzer.inspection.MtlogBatchInspection"
             runForWholeFile="true"
             displayName="mtlog template validation"
             groupName="Go"
@@ -26,24 +26,24 @@
             shortName="MtlogTemplateValidation">
             <quickFix order="first" displayName="Convert to PascalCase" 
                      familyName="mtlog quick fixes"
-                     implementationClass="com.mtlog.goland.quickfix.PascalCaseQuickFix"/>
+                     implementationClass="com.mtlog.analyzer.quickfix.PascalCaseQuickFix"/>
             <quickFix order="first" displayName="Fix template arguments" 
                      familyName="mtlog quick fixes"
-                     implementationClass="com.mtlog.goland.quickfix.TemplateArgumentQuickFix"/>
+                     implementationClass="com.mtlog.analyzer.quickfix.TemplateArgumentQuickFix"/>
         </localInspection>
         
         <projectService 
-            serviceImplementation="com.mtlog.goland.service.MtlogProjectService"/>
+            serviceImplementation="com.mtlog.analyzer.service.MtlogProjectService"/>
             
         <projectConfigurable
             parentId="tools"
-            instance="com.mtlog.goland.settings.MtlogSettingsConfigurable"
+            instance="com.mtlog.analyzer.settings.MtlogSettingsConfigurable"
             id="mtlog.settings"
             displayName="mtlog-analyzer"/>
             
         <advancedSetting 
             id="mtlog.debounce.delay"
-            service="com.mtlog.goland.settings.MtlogAdvancedSettings"
+            service="com.mtlog.analyzer.settings.MtlogAdvancedSettings"
             default="300"/>
             
         <notificationGroup id="mtlog.notifications" displayType="BALLOON"/>
@@ -51,7 +51,7 @@
     
     <actions>
         <action id="mtlog.analyzer.install"
-                class="com.mtlog.goland.actions.InstallAnalyzerAction"
+                class="com.mtlog.analyzer.actions.InstallAnalyzerAction"
                 text="Install mtlog-analyzer"
                 description="Install mtlog-analyzer using go install">
             <add-to-group group-id="ToolsMenu" anchor="last"/>
