+++ goland-plugin/src/test/kotlin/com/mtlog/analyzer/integration/MtlogSettingsIntegrationTest.kt
@@ -1,12 +1,12 @@
-package com.mtlog.goland.integration
+package com.mtlog.analyzer.integration
 
 import com.intellij.openapi.components.service
 import com.intellij.openapi.options.ShowSettingsUtil
 import com.intellij.testFramework.LoggedErrorProcessor
-import com.mtlog.goland.service.MtlogProjectService
-import com.mtlog.goland.settings.MtlogSettingsConfigurable
-import com.mtlog.goland.settings.MtlogSettingsState
-import com.mtlog.goland.service.AnalyzerDiagnostic
+import com.mtlog.analyzer.service.MtlogProjectService
+import com.mtlog.analyzer.settings.MtlogSettingsConfigurable
+import com.mtlog.analyzer.settings.MtlogSettingsState
+import com.mtlog.analyzer.service.AnalyzerDiagnostic
 import java.io.File
 
 class MtlogSettingsIntegrationTest : MtlogIntegrationTestBase() {
